Separate LiveView widget to a separate file, add some fits bits of applying background blur settings, fix compilation
