+++ lib/views/custom_widgets/wrappers/live_view_background.dart
@@ -1,10 +1,12 @@
-import 'dart:ui' as ui;
+import 'dart:ui';
 
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:momento_booth/managers/_all.dart';
+import 'package:momento_booth/models/settings.dart';
 import 'package:momento_booth/views/base/stateless_widget_base.dart';
+import 'package:momento_booth/views/custom_widgets/wrappers/live_view.dart';
 
 class LiveViewBackground extends StatelessWidgetBase {
 
@@ -16,7 +18,10 @@ class LiveViewBackground extends StatelessWidgetBase {
   });
 
   bool get _showLiveViewBackground => PhotosManager.instance.showLiveViewBackground;
+  BackgroundBlur get _backgroundBlur => SettingsManager.instance.settings.ui.backgroundBlur;
   LiveViewState get _liveViewState => LiveViewManager.instance.liveViewState;
+  int? get _textureIdMain => LiveViewManager.instance.textureIdMain;
+  int? get _textureIdBlur => LiveViewManager.instance.textureIdBlur;
 
   @override
   Widget build(BuildContext context) {
@@ -42,7 +47,7 @@ class LiveViewBackground extends StatelessWidgetBase {
             for (InfoBar notification in NotificationsManager.instance.notifications) ...[
               notification,
               const SizedBox(height: 8),
-            ]
+            ],
           ],
         ),
       ),
@@ -90,57 +95,29 @@ class LiveViewBackground extends StatelessWidgetBase {
       fit: StackFit.expand,
       children: [
         ColoredBox(color: Colors.green),
-        LiveView(
-          fit: BoxFit.cover,
-          textureId: LiveViewManager.instance.textureIdBlur,
-        ),
+        if (_backgroundBlur == BackgroundBlur.softwareResize)
+          LiveView(
+            fit: BoxFit.cover,
+            textureId: _textureIdBlur,
+          ),
+        if (_backgroundBlur == BackgroundBlur.textureBlur)
+          ImageFiltered(
+            imageFilter: ImageFilter.blur(sigmaX: 8, sigmaY: 8),
+            child: LiveView(
+              fit: BoxFit.cover,
+              textureId: _textureIdMain,
+            ),
+          ),
         AnimatedOpacity(
           duration: const Duration(milliseconds: 300),
           opacity: _showLiveViewBackground ? 1 : 0,
           curve: Curves.ease,
           child: LiveView(
             fit: BoxFit.contain,
-            textureId: LiveViewManager.instance.textureIdMain,
+            textureId: _textureIdMain,
           ),
         ),
       ],
-    );
-  }
-
-}
-
-class LiveView extends StatelessWidgetBase {
-
-  final BoxFit fit;
-  final int? textureId;
-
-  const LiveView({
-    super.key,
-    required this.fit,
-    this.textureId,
-  });
-
-  ui.FilterQuality get _filterQuality => SettingsManager.instance.settings.ui.liveViewFilterQuality.toUiFilterQuality();
-  double get _aspectRatio => SettingsManager.instance.settings.hardware.liveViewAndCaptureAspectRatio;
-
-  @override
-  Widget build(BuildContext context) {
-    return FittedBox(
-      fit: fit,
-      child: Observer(
-        builder: (_) {
-          return SizedBox(
-              width: _aspectRatio,
-              height: 1,
-              child: textureId != null
-                  ? Texture(
-                    textureId: textureId!,
-                    filterQuality: _filterQuality,
-                  )
-                  : null,
-            );
-        }
-      ),
     );
   }
